Change to ignore weekend, show day in report
Former-commit-id: 4735e6af844564ea3270acf81fba71e91d060dea

+++ src/main/java/com/tw/jiracalc/beans/CycleTimeBean.java
@@ -3,13 +3,13 @@ package com.tw.jiracalc.beans;
 import java.util.Map;
 
 public class CycleTimeBean {
-    private Map<String, Long> cycleTime;
+    private Map<String, Float> cycleTime;
 
-    public Map<String, Long> getCycleTime() {
+    public Map<String, Float> getCycleTime() {
         return cycleTime;
     }
 
-    public void setCycleTime(Map<String, Long> cycleTime) {
+    public void setCycleTime(Map<String, Float> cycleTime) {
         this.cycleTime = cycleTime;
     }
 }
